+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-05T05:06:33.389Z")
+    created: Optional[datetime] = parse_datetime("2026-02-05T09:59:21.290Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")